+++ connect/amazon/sqs/fake/src/main/kotlin/org/http4k/connect/amazon/sqs/FakeSQS.kt
@@ -32,6 +32,7 @@ class FakeSQS(
 
     override val app = routes(
         "/" bind POST to routes(
+            api.changeMessageVisibility(queues),
             api.deleteMessage(queues),
             api.deleteQueue(queues, deduplication, queueConfig),
             api.deleteMessageBatch(queues),